dev state: fixed most print string problems, when only first letter prints
git-svn-id: https://ja2svn.mooo.com/source/ja2@990 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ branches/Lesh/lpja2/src/strategic/campaign_init.cpp
@@ -90,7 +90,7 @@ altSectorsStartElementHandle(void *userData, const char *name, const char **atts
 				if ( isalpha(atts[1][0]) )
 					pData->uiRowNumber = (UINT32)(atts[1][0] - 0x40);
 				else
-					pData->uiRowNumber = atol(atts[1]);
+					pData->uiRowNumber = atoi(atts[1]);
 			}
 			pData->maxReadDepth++; //we are not skipping this element
 		}
